refact: #168 딕셔너리 생성
